REFACTOR: Restructure Project Modules and Improve Code Organization
- Move mappers, strategies, and async processing components to appropriate domain-specific packages
- Reorganize rate limiting annotations and components into a dedicated ratelimit package
- Remove deprecated and unused classes from global shared packages
- Update import statements to reflect new package structures
- Simplify base entity inheritance and remove redundant base classes
- Clean up event and async processing interfaces and implementations
