+++ src/main/java/com/example/lettuce/domain/auth/service/AuthService.java
@@ -19,12 +19,12 @@ import com.example.lettuce.domain.user.dto.UserInfo;
 import com.example.lettuce.domain.user.entity.Profile;
 import com.example.lettuce.domain.user.entity.User;
 import com.example.lettuce.domain.user.enums.UserRole;
+import com.example.lettuce.domain.user.mapper.ProfileMapper;
+import com.example.lettuce.domain.user.mapper.UserMapper;
 import com.example.lettuce.domain.user.service.ProfileMapperFactory;
 import com.example.lettuce.domain.user.service.UserService;
 import com.example.lettuce.global.framework.security.provider.JwtTokenProvider;
 import com.example.lettuce.global.shared.constant.AuthConstants;
-import com.example.lettuce.global.shared.mapper.ProfileMapper;
-import com.example.lettuce.global.shared.mapper.UserMapper;
 import com.example.lettuce.global.shared.exception.AuthenticationException;
 import com.example.lettuce.global.shared.exception.BaseException;
 import com.example.lettuce.global.shared.exception.code.ErrorCode;
