+++ src/main/java/com/example/lettuce/domain/carbonfootprint/repository/AsyncCarbonFootprintProductRepository.java
@@ -15,6 +15,7 @@ import org.springframework.stereotype.Repository;
 
 import com.example.lettuce.domain.carbonfootprint.dto.response.CarbonFootprintProductResponse;
 import com.example.lettuce.domain.carbonfootprint.service.CarbonFootPrintProduct;
+import com.example.lettuce.global.shared.async.AsyncMultiProcessor;
 import com.example.lettuce.global.shared.exception.BaseException;
 import com.example.lettuce.global.shared.exception.code.ErrorCode;
