--- src/main/java/com/example/lettuce/domain/carbonfootprint/repository/CarbonFootPrintProductResponse.java
@@ -1,5 +1,0 @@
-package com.example.lettuce.domain.carbonfootprint.repository;
-
-public class CarbonFootPrintProductResponse {
-
-}
