+++ src/main/java/com/example/lettuce/domain/carbonfootprint/repository/CarbonFootprintRewardRepository.java
@@ -5,5 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.lettuce.domain.carbonfootprint.entity.CarbonFootPrintReward;
 
 public interface CarbonFootprintRewardRepository extends JpaRepository<CarbonFootPrintReward, Long> {
-
 }
