--- src/main/java/com/example/lettuce/domain/user/service/EventProducer.java
@@ -1,5 +1,0 @@
-package com.example.lettuce.domain.user.service;
-
-public class EventProducer<T> {
-
-}
