+++ src/main/java/com/example/lettuce/domain/user/service/ProfileService.java
@@ -10,8 +10,9 @@ import com.example.lettuce.domain.user.dto.request.UpdateUserProfileRequest;
 import com.example.lettuce.domain.user.dto.response.ProfileResponse;
 import com.example.lettuce.domain.user.entity.Profile;
 import com.example.lettuce.domain.user.entity.User;
+import com.example.lettuce.domain.user.mapper.ProfileMapper;
 import com.example.lettuce.domain.user.repository.UserRepository;
-import com.example.lettuce.global.shared.mapper.ProfileMapper;
+import com.example.lettuce.domain.user.stragies.ProfileStrategy;
 import com.example.lettuce.global.shared.s3.S3Service;
 import com.example.lettuce.global.shared.s3.UploadImageInfo;
