+++ src/main/java/com/example/lettuce/global/shared/async/AsyncEventConsumer.java
@@ -1,4 +1,5 @@
-package com.example.lettuce.global.shared.event;
+package com.example.lettuce.global.shared.async;
+
 import java.util.List;
 
 public interface AsyncEventConsumer<T> {
