+++ src/main/java/com/example/lettuce/global/shared/async/AsyncEventProducer.java
@@ -1,4 +1,4 @@
-package com.example.lettuce.global.shared.event;
+package com.example.lettuce.global.shared.async;
 
 import java.util.List;
